requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
google-generativeai==0.5.4
torch==2.3.0
numpy==1.26.4
Pillow==10.3.0
scikit-learn==1.4.2
python-multipart==0.0.9
